requirements.txt

tensorflow
streamlit
opencv-python
pillow
numpy
